fix: persistent startup not loading roles (#86)
Roles are always calculated at startup in Clean mode, but in Persistent mode the calculation is skipped entirely, trusting the DB. The build process deleted part of the DB leading to no role shown. This post build script will ensure the integrity of the DB after the build process is completed. Add overwrite db option to enable fresh downloads